Adding a test missed in earlier commits of rsyslog test cases.
This test, sets up all the nodes to send syslog messages to
a single collector and verifies, syslog messages are written
into the db poperly with the node name tags as expected.

This test follows a static connection of rsyslog clients
to a single collector node. The first collector is picked
up in case the setup has multiple collector nodes.

Change-Id: Ib143a8ca03ba779c0a2a7664aa052d6e1781753c