Reducing bit-size of subnet mask from 30, to allow more hosts.
With recent changes in DNS, one more IP address from the bottom
line of a subnet would be picked up into the reserved pool.

So, with .0 being a reserved address and .1, .2 and .3 being
picked as contrail reserved address for default gateway, DNS
etc. the 30 bit mask does not leave any room for spawning VM
and assigning IP addresses to them.

This brings in the need to change the scripts which use 30 bit
mask to allow more IP addresses for the VM by reducing the bit
mask.

Change-Id: I87c52227a518f2a3d3a075cd014c3f25d974800c