Change in supervisor-vrouter restart routine.
The sup_vrouter_process_restart() routine currently waits
for 5 sec for the service to come up after restart.
The vrouter agent takes longer to sync routes and configs
from the control node after a restart. It takes around 70
sec. So, this change looks for vrouter-agent to be in active
state after a supervisor-vrouter restart.

Change-Id: I4001ab18efaa55a6427041fb1ab34917b1f185f5