Adding clean-up routine to test_max_vm_flows.
test_max_vm_flows changes the values for flow_cache_timeout
and vm_max_flows in the vrouter agent config file. Till now
this was being restored at the end of the test, which does not
take effect if the test fails and thus is a wrong way of cleanup.

This change registers a clean-up routine with the test infra
to take care that the default values are restored on test-completion
or test-failure, thereby not affecting other scripts with the
changed configs of vrouter.

Change-Id: I676a915c3a639032ff10170646084f98f690efc0