Adding test case to cover src & dst CIDR.
This test case added to the test_policy_acl script brings in test
coverage for source and destination CIDR address configuration
in a policy. The following tests are being added in this test-case.

1. Deny ICMP traffic between VM11 and VM12 where both are in VN1.
2. Allow ICMP traffic between VM11 and VM21 where VM11 is in VN1
and VM21 is in VN2.

Change-Id: I89d426a3b48d2d72225af7ebf609aff9f3ce72f1

Using get_cidrs() from VN fixture.

As per the comment in review 6918, to get the VN subnet
in cidr format, we should be using get_cidrs routine from
VN fixture.

Change-Id: I89d426a3b48d2d72225af7ebf609aff9f3ce72f1