Adding intial code for tor scale automation
Adding intial code for tor scale automation

Purpose:
Framework to configure scale config for TOR-BMS automation. This will
use exting library but not the fixture. Purpose is to load the scale config
as quickly as possible with minimal exiting framework overhead.

How to run:
1. Setup python path and create sanity_testbed.json as usual.
2. Create a a file under tor_scale folder with name tor_params.ini.
This file defines the scale config details.
3. Run the script load_config_parallel.py. This will load config
for all the TOR defined in configuration file in parallel
Change-Id: I73e9c70bd595fd88692a5d7d947967471ec1a63d