Fix for nova port delete.
    Before kilo release, nova would send port delete
explicitly in VM delete sequence, this would results in
deleting port specific objects across the system and network
delete goes through fine.
    But in Kilo, nova design has changed. Nova sends port
delete only for ports that it has created. If user has
created a port and attach a VM to it, nova will not send
port delete as a part of VM delete sequence. In Some test
sequence, port delete is not called in cleanup sequence
if port is created by test. This will keep port specific
backrefs in api-server, and VN delete fails in test clean
up sequence.

ClosesBug:#1498131

Change-Id: I91b577927e33a538985362f12f401f4678cece1d