Adding infra support for DPDK enabled compute
Objective is to run exting automated cases on a dpdk enabled
setup. In the setup few of the compute can be dpdk where rest
are running standard vrouter.

In a dpdk enabled compute to launch vm, we nedd hw:mem_page_size
set on flavor type as extra_specs.

To handle both standard and dpdk enabled compute, creating a new
flavour for dpdk setup.

Following logic is followed for selcting flavor

* No change if DPDK not enabled in setup.

* If DPDK enabled:

** node_name is Disabled: Leave it upon nova schduler to decide
   right compute with standard flavor. If flavor is not dpdk,
   nova scheduler will launch vm on stndard vrouter compute

** If node_name is some compute which is not dpdk enabled:  Use
   standard flavor.

** If node_name is some compute which is dpdk enabled: Use dpdk flavor.

Change-Id: Ib3c71fb58ea177da5eb1b62474c65cb87ec90484