Verifies that updating VM IP is not allowed.
Earlier, when we try to update the IP of the VM, it was possible
using neutron port-update. The issue in this was that Old IP
address also remained. So, there were 2 IP address in a tap interface.
As a fix, updating the VM IP is dissalowed.
This change checks the fix of issue1526260.

Change-Id: I4970102822b1c4cac7097c09992daba92c515f10