Test case to verify route leaking when a specific protocol mentioned
The change includes a test case in which verification of route
leaking is done when protocol field is set to any specific protocol
other than having a value as "any".
In this test case, route leaking is performed by rule with protocol
value as "tcp". Both failure and pass scenario for ICMP and TCP
respectively are tested.

Change-Id: Ic53498a611c51ad90b7ce30a97b81004050bfef8