Local VDNS record to be searched in discovery assigned DNS server
Currently, default DNS records for VMs are being searched in
first control node. This is not a right approach as records are
saved in control node assigned by discovery server.
Thus, we need to search for associated DNS server with that
agent and then search for records in that Control node.
Chnage has been made in multiple test cases.

Change-Id: I709d8b8aca271e0ab8b3faf9aa21a9c95f6be93b
Closes-Bug: 1579740