Verifies that updating VM IP is not allowed.
Earlier, when we try to update the IP of the VM, it was possible
using neutron port-update. The issue in this was that Old IP
address also remained. So, there were 2 IP address in a tap interface.
As a fix, updating the VM IP is dissalowed.
This change checks the fix of issue1526260.

It includes the changes after review comments of review
https://review.opencontrail.org/#/c/19521/

Change-Id: Ib4e328470991182521f6e1cd52b7e8ad9caeb6e6