Changing local command execution to cfgm node.
Local command execution to check for expected configurations in
rsyslog file fails because we are running the test from a docker
instance rather than a test node. So changing the commands to be
executed on the cfgm node where the rsyslog.conf file is expected
to be populated with the desired parameters.

Change-Id: I90134a9f36106414e99c5b839e19c6769eee127e