Done modification in R3.0 VDNS code for compatibility with mainline
A proc was modified earlier but due to mismatching mainline and R3.0
code, the change was missed in one of the file in R3.0.
Apart from this, the base file in serial scripts is deleted completely.
This is also done to make it compatible with mainline and new test
infra. Proper import statement include in the test suite.

Change-Id: I264bf65574451b6e9ca3abe0805cbef4e7d9152e