VDNS test code cleanup + A Sanity Fix
File named vdns_test.py was having a lot of duplicate code. At
some time in the past, the code in this file was modified as
per new infra changes and distributed between test_vdns.py and
base.py

Thus, deleting "vdns_test.py" .
Also, as some code in this file is not duplicate, adding and modifying
the unique code as per new infra in the file "test_vdns.py"

Sanity Fix: Apart from this, Adding a sleep of 3 seconds in a
test case to give time for records to get updated in VMs.
Sleep added in 1 second polls

Change-Id: I8cbfc0c16ffc1daccdee207548026d9ed8575552