Dynamically identified the sport value of flow
Earlier, the sport value was hard coded to 0.
Recently, some change was done in ICMP traffic generation through
scapy to distinguish 2 different ping streams.
A header ID was introduced which randomize the sport of flow.
Thus, the harcoded value of '0' can no longer be used to capture
a flow. It has to be dynamically identified.

Change-Id: I5614258d033d044b074b0b5d8c329eb5a8c515d2