Optimization for vcenter gateway
Earlier the port was being created after the vm was boot up.
This used to cause dalay as ,while booting up,vm used to spend
significant time in trying dhcp, but ultimately in vain.

We should create the port just after launching the vm in vcenter,
so that the dhcp does not fail during vm boot up.

Change-Id: I5437d1c69f1d5b1ed977bc46805d58752fba1d66