Moving the pbb evpn test cases to serial_scripts
PBB EVPN is supported in MPLSoUDP and MPLSoGRE encapsulations
and VXLAN is not supported. Majority testbeds are provisioned
with default encap priority as: VXLAN, MPLSoUDP, MPLSoGRE. So,
to run pbb evpn test cases, in the tests now we are changing the
encap priority to: MPLSoUDP, MPLSoGRE and VXLAN. Also, moving the
test cases to serial_scripts so that this setting will not
effect the parallel execution of other scripts.

Change-Id: I4a7a0ace3b5ca37ccc4701460ac19ec7150d001b