Skipping IPv6 cases in vcenter environment
Skip ipv6 cases in vcenter environment for policy and
ecmp.
Enabled ecmp cases for vrouter gateway mode.

Fixed vcenter test cases where interface/network for the
vm changed.The contrail-test-ci changes would go in a separate
review.We need to refresh the interface/network information
in the vm fixture.

We are passing the refresh parameter from test and updating
the vm fixture object based on whether its True or False

Change-Id: I6959b431c8403c356eb6fbbc819708df12fea838