Moving mvpn scripts to serial_scripts folder
As of now, mvpn can not be enabled through VNC API. control
docker entrypoint.sh file needs to be added with
'mvpn_ipv4_enable=1' knob and restart of control docker is
required. This step can impact the other tests when it is run
in parallel. So, moving the scripts to serial_scripts folder
as of now. Once, VNC API support is present, we can move these
scripts to scripts directory from serial_scripts directory.

Change-Id: Ib7184f82d210f843c6c8d729a13bce50fecc7785
Closes-Bug: 1764228