Changing the logic to get the VRF_ID based on both the vn_fq_name as well as the SI name
In cases where multiple SIs, esp. transparent ones are found on the same
compute, the check only for vn_fq_name to get the vrf_id could be
incorrect.

Hence changing the logic to get the VRF_ID based on both the vn_fq_name
as well as the SI name.

Change-Id: Ica8c4901eea8493594c11f765e43d5a744c5bf0d
Closes-Bug: #1794561