Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

We need parametr to pass default contrail domain to few specific places in contrail for ensure test to work on JuJu. In order for the tests to work, we must by default set this domain parameter to default-domain for all cases except JuJu. We already have two parameters stack_domain and admin_domain in contrail_test_init, but if we set them to default-domain by default then the tests do not pass, since the default domain in OpenStack is not "default-domain" but «default". "default-domain" is the default domain for contrail. Therefore, to set the default contrail domain (not openstack) only for individual contrail parameters, we cannot use the existing stack_domain or admin_domain params in contrail_init_tests and are forced to add a new parameter.

Patch-set: 1
Reviewer: Gerrit User 872 <872@c72cd27d-1743-4a55-84a4-cd13b8400c52>