use newer pip2 to install packages
pinning setuptools as <47.0 is not so correct even if it can be installed.
here https://pypi.org/project/setuptools/46.4.0/ we can see that it
supports only python3 also.

but is in old pip - it doesn't read markers from pypi and can
install packages marked for python3 into python2 env.

Related-Jira-Bug: CEM-16113
Change-Id: I59c1b969441c44028ac984596bea44f7f273282b